FIX: impor/export YAML configuration (bad way)
This is an ugly way to fix it. When dumping YAML to file, one should stay
consistent and not mix up addict.Dict() w/ dict(). As this seems to happen here,
the quickiest solution was to convert the whole dict to Dict() and call .to_dict()
to ensure any subkey is now exposing a real dict() before dumping. Instead, one should
remain consistent when pushing data into config/profile