BANK: Save into bank when a run completes
It may also be a possibility to incrementally save results in the bank, allowing
post-exec analysis to be run while the test-suite is still running. To do so,
a Bank allow to dump results directly as soon as the JSON can be properly formatted.
This commit only call save_from_buildir() once the test-suite is complete