ORCH: fix bad scheduling
As jobs picked as deps were not removed from the manager
(for efficiency purpose), such jobs might be picked again by
another set to be scheduled again, causing the test-base to
grow significantly.

Also, the launch system changed. Now as much 'runners' as possible
are created (max one per allowed concurrent launch) and queues are
used to communicate sets:
  - One is used for manager -> runners communication (ready_q)
  - One is used to publish result to manager (completed_q)

Still, one bug remains here when some Ctrl-C is received, *probably*
because of Click registering an atexit() function, but the function
is not able to access a valid context (multithread situation)